requirements.txt
# --- Base FastAPI ---
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary
pydantic>=2.7.0
passlib[bcrypt]

bcrypt>=3.2.2
python-jose[cryptography]


langchain-core>=0.2.10
langchain>=0.2.6
langchain-community>=0.2.6


langchain-chroma>=0.1.1
langchain-google-genai>=1.0.6
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.2.1

sentence-transformers>=2.2.2
scikit-learn



mlflow==2.14.1
boto3   

pytest
httpx